Fixed some bugs with mutated trees
Realized that `expand_tree` and `contract_tree` mutate input tree.
This was the reason some of the tests worked, so I had to work on the
tests and add some options.

Remember that python is pass by assignment.
Mutable things will be mutable after passing, immutable will be
immutable.
A `nltk.Tree` is mutable. Immutable are e.g. integers, strings, ...
Other mutables are lists, (dictionaries?)